backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.8
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.354
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pip==22.3.1
pydantic==1.10.9
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
setuptools==65.5.0
SQLAlchemy==2.0.16
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.6.3
typing-inspect==0.9.0
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
pandas==2.1.4
psycopg2-binary==2.9.9